+++ packages/forma-36-website/src/components/Layout.js
@@ -4,7 +4,9 @@ import Helmet from 'react-helmet';
 import { useStaticQuery, graphql } from 'gatsby';
 import '@contentful/f36-components/dist/styles.css';
 import '@contentful/forma-36-fcss/dist/styles.css';
-import { css } from 'emotion';
+
+import { css } from '@emotion/core';
+
 import Header from './Header';
 import Promo from './Promo';
 import Container from './Container';
@@ -14,13 +16,10 @@ import './Layout.css';
 const styles = {
   main: css`
     display: flex;
-    flex: 2;
+    height: calc(100vh - 70px);
   `,
-  test: css`
-    display: flex;
-    flex-direction: column;
-    height: 100vh;
-    overflow-y: hidden;
+  withPromo: css`
+    height: calc(100vh - 107px);
   `,
 };
 
@@ -51,8 +50,10 @@ const Layout = (props) => {
     }
   `);
 
+  const withPromo = !!data.site.siteMetadata.promoText;
+
   return (
-    <div className={styles.test}>
+    <>
       <Helmet
         title={data.site.siteMetadata.title}
         meta={[
@@ -69,7 +70,7 @@ const Layout = (props) => {
         <html lang="en" />
       </Helmet>
 
-      {data.site.siteMetadata.promoText && (
+      {withPromo && (
         <Promo
           text={data.site.siteMetadata.promoText}
           linkHref={data.site.siteMetadata.promoLink}
@@ -79,8 +80,7 @@ const Layout = (props) => {
       )}
 
       <Header />
-
-      <div className={styles.main}>
+      <div css={[styles.main, withPromo && styles.withPromo]}>
         <Navigation
           menuItems={data.site.siteMetadata && data.site.siteMetadata.menuLinks}
           currentPath={props && props.location && props.location.pathname}
@@ -93,13 +93,13 @@ const Layout = (props) => {
           {props.children}
         </Container>
       </div>
-    </div>
+    </>
   );
 };
 
 Layout.propTypes = {
   children: PropTypes.node,
-  location: PropTypes.object.isRequired,
+  location: PropTypes.object,
 };
 
 export default Layout;