+++ packages/forma-36-react-components/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.95.1](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-components@3.95.0...@contentful/forma-36-react-components@3.95.1) (2021-08-04)
+
+
+### Bug Fixes
+
+* **Icon:** add types for otherProps ([#1150](https://github.com/contentful/forma-36/issues/1150)) ([eb8ddb5](https://github.com/contentful/forma-36/commit/eb8ddb5d96eb5d59f6fce527eb9c1ce09d9b0fca))
+
+
+
+
+
+# [3.95.0](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-components@3.94.2...@contentful/forma-36-react-components@3.95.0) (2021-08-04)
+
+
+### Features
+
+* **Puzzle:** use web app version of puzzle icon ([#1146](https://github.com/contentful/forma-36/issues/1146)) ([b76cd56](https://github.com/contentful/forma-36/commit/b76cd563e73ef0731d250c6acb2c0b4df357963f))
+
+
+
+
+
+## [3.94.2](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-components@3.94.1...@contentful/forma-36-react-components@3.94.2) (2021-08-04)
+
+
+### Bug Fixes
+
+* workbench sidebar label ([#1148](https://github.com/contentful/forma-36/issues/1148)) ([a2a72bb](https://github.com/contentful/forma-36/commit/a2a72bbcf2ca3c48945284c7838ebbf348f3a182))
+
+
+
+
+
 ## [3.94.1](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-components@3.94.0...@contentful/forma-36-react-components@3.94.1) (2021-08-03)